prevent segfault when passing matrix data
Previously, if M or C matrices (sparse or dense) were passed without a
diagonal, a segfault would occur. Now, this is (partially) prevented. If
the user uses the method `setMatrixData`, this is completely prevented.
If the user uses the function `setSparseMatrixData`, it is assumed that
no diagonal elements have been set.

This change also adds some tests